Add 1.9 build tag for type alias parser
